Refactors User and Cart controllers.
create function (and all following CRUD) for
Cart transferred to carts controller
('/controllers/carts.js')
