+++ app/controllers/users.js
@@ -7,7 +7,7 @@ const models = require('app/models');
 const User = models.user;
 
 
-const lineItems = models.lineItem;
+// const lineItems = models.lineItem;
 
 
 const crypto = require('crypto');
@@ -114,23 +114,23 @@ const changepw = (req, res, next) => {
   ).catch(makeErrorHandler(res, next));
 };
 
-const addToCart = (req, res, next) => {
-  // console.log(req.body)
-  // let user = Object.assign(req.body.lineItems, {
-  //   // _owner: req.currentUser._id
-  // });
-  // lineItems.push(req.body.cart)
-  // .then(lineItems => res.json({ lineItems }))
-  // .catch(err => next(err));
-  let lineItem = req.body.lineItems;
-    console.log(lineItem);
-  User.findById(req.currentUser._id)
-  .then((user) => user.update({$push: {"cart": lineItem}}))
-  .then(() => res.sendStatus(200))
-  .catch(err => next(err));
-
-  // return user.update({$push: {"cart": lineItem}})
-};
+// const addToCart = (req, res, next) => {
+//   // console.log(req.body)
+//   // let user = Object.assign(req.body.lineItems, {
+//   //   // _owner: req.currentUser._id
+//   // });
+//   // lineItems.push(req.body.cart)
+//   // .then(lineItems => res.json({ lineItems }))
+//   // .catch(err => next(err));
+//   let lineItem = req.body.lineItems;
+//     console.log(lineItem);
+//   User.findById(req.currentUser._id)
+//   .then((user) => user.update({$push: {"cart": lineItem}}))
+//   .then(() => res.sendStatus(200))
+//   .catch(err => next(err));
+//
+//   // return user.update({$push: {"cart": lineItem}})
+// };
 
 // const create = (req, res, next) => {
 //   let example = Object.assign(req.body.example, {
@@ -150,7 +150,7 @@ module.exports = controller({
   signin,
   signout,
   changepw,
-  addToCart
+  // addToCart
 }, { before: [
   { method: authenticate, except: ['signup', 'signin'] },
 ], });
