+++ config/routes.js
@@ -15,7 +15,7 @@ module.exports = require('lib/wiring/routes')
 .post('/sign-in', 'users#signin')
 .delete('/sign-out/:id', 'users#signout')
 .patch('/change-password/:id', 'users#changepw')
-.post('/additem', 'users#addToCart')
+.post('/additem', 'carts#addToCart')
 
 
 .resources('users', { only: ['index', 'show'] })
@@ -30,8 +30,8 @@ module.exports = require('lib/wiring/routes')
 .resources('products')
 
 .get('/carts', 'carts#index')
-.get('/carts/:id', 'carts#show')
-.post('/carts', 'carts#create')
+// .get('/carts/:id', 'carts#show')
+.post('/carts', 'carts#addToCart')
 
 
 // all routes created
