+++ scripts/addToCart.sh
@@ -1,5 +1,5 @@
 ID=57389ed4f26cbe6d1075b42f
-TOKEN=b+RRi6FtwSLec7ryTa9+eJLpqVeyZ7EmSrTCCpKzDMQ=--K063yR6PavKLfTZZAK0Z7YOCvJ4uBC1n03oDk8ywW3M=
+TOKEN=bdLLzCxmF5KIfqot2cjyZeaOJVRRxXwIjYuH3mupnow=--CicJ6/X8dxY3eipui+AiP8G1BNW2fxVz6wTEOgEAEVc=
 curl --include --request POST http://localhost:3000/additem \
   --header "Authorization: Token token=$TOKEN" \
   --header "Content-Type: application/json" \
