+++ scripts/orders-create.sh
@@ -1,9 +1,8 @@
 TOKEN="2a4CTcSPtFvk8imstqtqmayTZOGtKtnpp10z8o0TTMs=--AH7GSLohUWDOvjgHr4mIWhZyZk98DVv8gxg/jbcmZ+I="
 curl --include --request POST http://localhost:3000/orders \
-  --header "Authorization: Token token=DnBdbZ8ZG+i39Rz6244rFy4YoEnDJ2vtFljq1YsNGu8=--Epl+j/9TQ3n3YwOuhYPGwgXqI+BCbhsUOMDA4egoJs4=" \
+  --header "Authorization: Token token=nqDPgVuDMHsMWG09PRMmoX3lx2QarC6GPr669s36QUI=--fTbpOwg+b1JGuMkSiTHokabkMx9tRzIYEP+yILJbZxY=" \
   --header "Content-Type: application/json" \
   --data '{
-    "_owner": "57389ed4f26cbe6d1075b42f",
     "lineItems": [{
         "product": {
           "quantity": "1",
@@ -11,7 +10,6 @@ curl --include --request POST http://localhost:3000/orders \
           "name": "The Art of the Deal",
           "price": "1000000000"
         }
-      }]
   }'
